emotional vector values all between -1 and 1
First we map all values between 0 and 2 and use that to average
because negative values were vausing problems.

I'm not sure if this is the best way to do this, will investigate
with more comments tomorrow.

I want to try by taking the absolute value of the polarity or
something along those lines.