Averaging based on absolute value now
Average emotion is based on all the concepts found in a comment
weighted by their polarity. The absolute value of the polarity is
taken so that emotions don't flip or cause math problems.

This causes emotions to always hover around 0 because polarity
will always be between 0 and 1 so the emotion value will get lower
when you multiply it by the polarity.

Next technique to try is to ignore polarity for now.